BookyNewEnv/Booky/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
dj-database-url==1.2.0
Django==4.1.4
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycryptodome==3.16.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.2.0
